pyproject.toml
[project]
name = "web-scrape-with-llm-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search==4.1.1",
    "langchain==0.0.352",
    "langchain-openai>=0.0.6",
    "matplotlib>=3.8.0",
    "openai>=1.6.1",
    "openpyxl>=3.1.2",
    "pandas>=2.1.4",
    "python-box>=7.1.1",
    "python-dotenv>=1.0.0",
    "wikipedia>=1.4.0",
    "xlsxwriter>=3.1.9",
]
